+++ src/backend/joanie/settings.py
@@ -241,6 +241,8 @@ class Base(Configuration):
             "django_filters.rest_framework.DjangoFilterBackend"
         ],
         "EXCEPTION_HANDLER": "joanie.core.api.exception_handler",
+        "DEFAULT_PAGINATION_CLASS": "rest_framework.pagination.PageNumberPagination",
+        "PAGE_SIZE": 20,
         "DEFAULT_VERSIONING_CLASS": "rest_framework.versioning.URLPathVersioning",
     }